fix(web-console): address market data review edge cases
zen-mcp-review: approved

continuation-id: direct-claude-approved-gemini-snippet-verified

+++ apps/market_data_service/schemas.py
@@ -5,6 +5,7 @@ Defines request/response models used by HTTP API routes.
 """
 
 from datetime import date, datetime
+from decimal import Decimal
 
 from pydantic import BaseModel, Field
 
@@ -44,8 +45,8 @@ class LatestQuoteResponse(BaseModel):
     """Latest top-of-book quote response payload."""
 
     symbol: str = Field(..., description="Stock symbol")
-    bid_price: float | None = Field(default=None, description="Best bid price")
-    ask_price: float | None = Field(default=None, description="Best ask price")
+    bid_price: Decimal | None = Field(default=None, description="Best bid price")
+    ask_price: Decimal | None = Field(default=None, description="Best ask price")
     bid_size: int | None = Field(default=None, description="Best bid size")
     ask_size: int | None = Field(default=None, description="Best ask size")
     timestamp: datetime | None = Field(default=None, description="Quote timestamp (UTC)")
